Use onActivityCreated in frag instead of onCreate since the activity's onCreate may not finish by the time its called for the frag -Re-adding the findById realm extension -Age is stored as a long on the server

+++ app/src/main/java/me/kmmiller/theduckypodcast/models/UserModel.kt
@@ -12,13 +12,13 @@ open class UserModel : RealmObject() {
     @PrimaryKey
     var id: String = ""
 
-    var age = 0
+    var age = 0L
     var gender = ""
     var state = ""
 
     fun toRealmModel(document: DocumentSnapshot) {
         id = document.id
-        age = document["age"] as? Int ?: 0
+        age = document["age"] as? Long ?: 0L
         gender = document["gender"] as? String ?: ""
         state = document["state"] as? String ?: ""
     }
